Fix release
