+++ benchmarks/FastGeoMesh.Benchmarks/Program.cs
@@ -147,22 +147,4 @@ namespace FastGeoMesh.Benchmarks
             Console.WriteLine("✅ Comprehensive validation completed!");
         }
     }
-
-    /// <summary>
-    /// Category filter for BenchmarkDotNet to run specific benchmark groups.
-    /// </summary>
-    public class CategoryFilter : IFilter
-    {
-        private readonly string _category;
-
-        public CategoryFilter(string category)
-        {
-            _category = category;
-        }
-
-        public bool Predicate(BenchmarkCase benchmarkCase)
-        {
-            return benchmarkCase.Descriptor.Categories.Contains(_category, StringComparer.OrdinalIgnoreCase);
-        }
-    }
 }
